pyproject.toml
[tool.poetry]
name = "pubscience"
version = "0.1.0"
description = ""
authors = ["Bram van Es <bramiozo@gmail.com>"]
license = "GPL-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
deepl = "^1.18.0"
google-cloud = "^0.34.0"
python-dotenv = "^1.0.1"
openai = "^1.44.1"
pandas = "^2.2.2"
py7zr = "^0.22.0"
google-cloud-translate = "^3.16.0"
google-generativeai = "^0.7.2"
pydantic = "^2.8.2"
anthropic = "^0.34.1"
ftfy = "^6.2.3"
python-benedict = "^0.33.2"
groq = "^0.11.0"
datasets = "^3.0.0"
accelerate = "^0.34.2"
sentencepiece = "^0.2.0"
nltk = "^3.9.1"
#torch = { version = "2.4.1", source = "torch" }
torch = "^2.4.1"
transformers = "^4.45.1"
zstandard = "^0.23.0"
pyreadstat = "^1.2.8"
sas7bdat = "^2.2.3"
ell-ai = "^0.0.14"
lxml = "^5.3.0"
pysbd = "^0.3.4"
bitsandbytes = "^0.45.0"
cachetools = "^5.5.0"
unsloth = "^2024.12.12"
sacremoses = "^0.1.1"

httpx = "0.27.2"
[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
seaborn = "^0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"
